backend/requirements.txt
alembic
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
psycopg[binary]
psycopg-pool
psycopg2-binary
python-jose[cryptography]
slowapi
pytest
httpx
python-multipart
aiofiles
Pillow
